+++ lib/arduino_ci/arduino_backend.rb
@@ -113,8 +113,15 @@ module ArduinoCI
     # @param boardname [String] The board to test
     # @return [bool] Whether the board is installed
     def board_installed?(boardname)
-      # capture_json("core", "list")[:json].find { |b| b["ID"] == boardname } # nope, this is for the family
       run_and_capture("board", "details", "--fqbn", boardname)[:success]
+    end
+
+    # check whether a board family is installed (e.g. arduino:avr)
+    #
+    # @param boardfamily_name [String] The board family to test
+    # @return [bool] Whether the board is installed
+    def boards_installed?(boardfamily_name)
+      capture_json("core", "list")[:json].any? { |b| b["ID"] == boardfamily_name }
     end
 
     # install a board by name
@@ -127,6 +134,15 @@ module ArduinoCI
         run_and_capture("core", "install", boardfamily, "--additional-urls", @additional_urls.join(","))
       end
       result[:success]
+    end
+
+    # Find out if a library is available
+    #
+    # @param name [String] the library name
+    # @return [bool] whether the library can be installed via the library manager
+    def library_available?(name)
+      # the --names flag limits the size of the response to just the name field
+      capture_json("lib", "search", "--names", name)[:json]["libraries"].any? { |l| l["name"] == name }
     end
 
     # @return [Hash] information about installed libraries via the CLI